# Backend/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Food Request API
  version: 1.0.0
paths:
  /food/{userId}:
    post:
      summary: Create a food listing with an image
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                foodRequestDTO:
                  type: string
                  description: JSON object as a string
                  example: |
                    {
                      "requestType": "Donation",
                      "foodItem": "Canned Beans",
                      "description": "A can of beans in good condition",
                      "quantity": 10,
                      "expiryDate": "2024-06-27T09:37:06.708Z",
                      "pickupLocation": "123 Charity St.",
                      "listingStatus": "Available"
                    }
                image:
                  type: string
                  format: binary
                  description: The image file to upload
                  example: "sample_image.png"
      responses:
        '201':
          description: Successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
                example: "Bad Request: Error message"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
                example: "Internal Server Error: Error message"
components:
  schemas:
    FoodRequestDTO:
      type: object
      properties:
        requestType:
          type: string
          example: "Donation"
        foodItem:
          type: string
          example: "Canned Beans"
        description:
          type: string
          example: "A can of beans in good condition"
        quantity:
          type: integer
          example: 10
        expiryDate:
          type: string
          format: date-time
          example: "2024-06-27T09:37:06.708Z"
        pickupLocation:
          type: string
          example: "123 Charity St."
        listingStatus:
          type: string
          example: "Available"